[JBEAP-1504, HAL-910] table focus changes after editting
